% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_grid_points_to_polygons.R
\name{attribute_grid_points_to_polygons}
\alias{attribute_grid_points_to_polygons}
\title{predict_grid}
\usage{
attribute_grid_points_to_polygons(prediction_grid, polygon_edges)
}
\arguments{
\item{prediction_grid}{test}

\item{polygon_edges}{test}
}
\description{
predict_grid
}
