% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bleiglas.R
\docType{package}
\name{bleiglas-package}
\alias{bleiglas}
\alias{bleiglas-package}
\title{bleiglas: Spatiotemporal Data Interpolation and Visualisation based on 3D Tessellation}
\description{
Employs Voro++ for the calculation of three dimensional Voronoi diagrams from
    input point clouds. This is a special form of tessellation where each polygon is defined 
    as the area closest to one particular seed point. Voronoi diagrams have useful applications
    in - among others - astronomy, material science or geography and bleiglas provides functions 
    to make 3D tessellation more readily available as a mean for data visualisation and interpolation.
    It can be used for any 3D point cloud, but the output is optimized for spatiotemporal applications 
    in archaeology.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nevrome/bleiglas}
  \item Report bugs at \url{https://github.com/nevrome/bleiglas/issues}
}

}
\author{
\strong{Maintainer}: Clemens Schmid \email{clemens@nevrome.de} (\href{https://orcid.org/0000-0003-3448-5715}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Stephan Schiffels (\href{https://orcid.org/0000-0002-1017-9150}{ORCID}) [contributor]
}

}
\keyword{internal}
