% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pred_grid.R
\name{create_pred_grid}
\alias{create_pred_grid}
\title{create_pred_grid}
\usage{
create_pred_grid(polygon_edges, res_x = 50000, res_y = 50000, res_z = 200)
}
\arguments{
\item{polygon_edges}{test}

\item{res_x}{test}

\item{res_y}{test}

\item{res_z}{test}
}
\description{
create_pred_grid
}
