% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_polygons.R
\name{cut_polygons}
\alias{cut_polygons}
\title{cut_polygons}
\usage{
cut_polygons(x, cuts)
}
\arguments{
\item{x}{data.table with output of voro++ as produced with
\link{tessellate} and then \link{read_polygon_edges}}

\item{cuts}{numeric vector with z-axis coordinates where cuts should be applied}
}
\value{
list of lists. One list element for each cutting surface and within these
data.tables for each 2D polygon that resulted from the cutting operation.
Each data.table holds the corner coordinates for one 2D polygon.
}
\description{
Figuratively cut horizontal slices of a 3D, tessellated cube.
}
\examples{
random_unique_points <- unique(data.table::data.table(
  id = NA,
  x = runif(10, 0, 100000),
  y = runif(10, 0, 100000),
  z = runif(10, 0, 100)
))
random_unique_points$id <- 1:nrow(random_unique_points)

voro_output <- tessellate(random_unique_points, unit_scaling = c(0.001, 0.001, 1))

polygon_points <- read_polygon_edges(voro_output)

cut_surfaces <- cut_polygons(polygon_points, c(20, 40, 60))

cut_surfaces_sf <- cut_polygons_to_sf(cut_surfaces, crs = 25832)
}
