% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_polygons_to_sf.R
\name{cut_polygons_to_sf}
\alias{cut_polygons_to_sf}
\title{cut_polygons_to_sf}
\usage{
cut_polygons_to_sf(x, crs)
}
\arguments{
\item{x}{data.table. Output of cut_polygons}

\item{crs}{coordinate reference system of the resulting 2D polygons}
}
\description{
cut_polygons_to_sf
}
