% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{line_segment_plane_intersection}
\alias{line_segment_plane_intersection}
\alias{line_segment_plane_intersection_multi}
\title{Find the intersection points of a line segment and a plane}
\usage{
line_segment_plane_intersection(point_a, point_b, plane_point, plane_normal)

line_segment_plane_intersection_multi(points, plane_point, plane_normal)
}
\arguments{
\item{point_a}{numeric vector. Coordinates of point A}

\item{point_b}{numeric vector. Coordinates of point B}

\item{plane_point}{numeric vector. Coordinates of plane point}

\item{plane_normal}{numeric vector. Plane normal vector}
}
\value{
...
}
\description{
\code{line_segment_plane_intersection} ...

Based on this solution:
\url{https://rosettacode.org/wiki/Find_the_intersection_of_a_line_with_a_plane#C.2B.2B}
}
\examples{
line_segment_plane_intersection(
  c(2, 2, 0),
  c(2, 2, 5),
  c(0, 0, 10),
  c(0, 0, 1)
) 

}
