% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pred_grid.R
\name{predict_grid}
\alias{predict_grid}
\title{predict_grid}
\usage{
predict_grid(polygon_edges, prediction_grid)
}
\arguments{
\item{polygon_edges}{test}

\item{prediction_grid}{test}
}
\description{
predict_grid
}
