% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_polygon_edges.R
\name{read_polygon_edges}
\alias{read_polygon_edges}
\title{read_polygon_edges}
\usage{
read_polygon_edges(x)
}
\arguments{
\item{x}{character vector with raw, linewise output of voro++ as produced with 
\link{tessellate} when \code{output_definition = "\%i*\%P*\%t"}}
}
\value{
data.frame with columns for the coordinates x, y and z of the starting and 
end point of each polygon edge
}
\description{
Special reader function for polygon edge output of voro++.
}
