% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_grid_points.R
\name{attribute_grid_points}
\alias{attribute_grid_points}
\title{predict_grid}
\usage{
attribute_grid_points(polygon_edges, prediction_grid)
}
\arguments{
\item{polygon_edges}{test}

\item{prediction_grid}{test}
}
\description{
predict_grid
}
