% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bleiglas.R
\docType{package}
\name{bleiglas-package}
\alias{bleiglas}
\alias{bleiglas-package}
\title{bleiglas: 3D Tessellation with Voro++}
\description{
Some helper functions for 3D tessellation with voro++ and subsequent horizontal cutting of the resulting polygons for 2D plotting.
}
\author{
\strong{Maintainer}: Clemens Schmid \email{clemens@nevrome.de} (\href{https://orcid.org/0000-0003-3448-5715}{ORCID}) [copyright holder]

}
\keyword{internal}
