% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bleiglas.R
\docType{package}
\name{bleiglas-package}
\alias{bleiglas}
\alias{bleiglas-package}
\title{bleiglas: 3D Tessellation with Voro++}
\description{
3D tessellation with voro++ and subsequent horizontal cutting of the resulting polygons for 2D plotting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nevrome/bleiglas}
  \item Report bugs at \url{https://github.com/nevrome/bleiglas/issues}
}

}
\author{
\strong{Maintainer}: Clemens Schmid \email{clemens@nevrome.de} (\href{https://orcid.org/0000-0003-3448-5715}{ORCID}) [copyright holder]

}
\keyword{internal}
