% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_polygons.R
\name{cut_polygons}
\alias{cut_polygons}
\title{cut_polygons}
\usage{
cut_polygons(x, cuts)
}
\arguments{
\item{x}{data.frame with output of voro++ as produced with 
\link{tessellate} and then \link{read_polygon_edges}}

\item{cuts}{numeric vector with z-axis coordinates where cuts should be applied}
}
\value{
simple features object with 2D polygons that result from the cutting operation
}
\description{
Figuratively cut horizontal slices of a 3D, tessellated cube.
}
