% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_polygons_to_sf.R
\name{cut_polygons_to_sf}
\alias{cut_polygons_to_sf}
\title{cut_polygons_to_sf}
\usage{
cut_polygons_to_sf(x, crs)
}
\arguments{
\item{x}{list of lists of data.tables. Output of cut_polygons}

\item{crs}{coordinate reference system of the resulting 2D polygons.
Integer with the EPSG code, or character with proj4string}
}
\value{
sf object with one row for each 2D polygon
}
\description{
Transform the polygon cut slices to the sf format. This only makes sense
if the x and y coordinate of your input dataset are spatial coordinates.
}
\examples{
random_unique_points <- unique(data.table::data.table(
  id = NA,
  x = runif(10, 0, 100000),
  y = runif(10, 0, 100000),
  z = runif(10, 0, 100)
))
random_unique_points$id <- 1:nrow(random_unique_points)

voro_output <- tessellate(random_unique_points, unit_scaling = c(0.001, 0.001, 1))

polygon_points <- read_polygon_edges(voro_output)

cut_surfaces <- cut_polygons(polygon_points, c(20, 40, 60))

cut_surfaces_sf <- cut_polygons_to_sf(cut_surfaces, crs = 25832)
}
