% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_grid.R
\name{predict_grid}
\alias{predict_grid}
\alias{attribute_grid_points_to_polygons}
\title{predict_grid}
\usage{
predict_grid(x, prediction_grid, ...)

attribute_grid_points_to_polygons(prediction_grid, polygon_edges)
}
\arguments{
\item{x}{test}

\item{prediction_grid}{test}

\item{...}{test}

\item{polygon_edges}{test}
}
\value{
test
}
\description{
predict_grid
}
