% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_polygon_edges.R
\name{read_polygon_edges}
\alias{read_polygon_edges}
\title{read_polygon_edges}
\usage{
read_polygon_edges(x)
}
\arguments{
\item{x}{character vector with raw, linewise output of voro++ as produced with 
\link{tessellate} when \code{output_definition = "\%i*\%P*\%t"}}
}
\value{
data.table with columns for the coordinates x, y and z of the starting and 
end point of each polygon edge
}
\description{
Special reader function for polygon edge output of voro++.
}
\examples{
random_unique_points <- unique(data.table::data.table(
  id = NA,
  x = runif(10),
  y = runif(10),
  z = runif(10)
))
random_unique_points$id <- 1:10

voro_output <- tessellate(random_unique_points)

polygon_points <- read_polygon_edges(voro_output)

cut_surfaces <- cut_polygons(polygon_points, c(0.2, 0.4, 0.6))

cut_surfaces_sf <- cut_polygons_to_sf(cut_surfaces, crs = 25832)

}
