% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_polygon_edges.R
\name{read_polygon_edges}
\alias{read_polygon_edges}
\title{read_polygon_edges}
\usage{
read_polygon_edges(x, rescale = TRUE)
}
\arguments{
\item{x}{character vector with raw, linewise output of voro++ as produced with
\link{tessellate} when \code{output_definition = "\%i*\%P*\%t"}}

\item{rescale}{Should the output of \link{tessellate} be back-rescaled according to its 
\code{unit_scaling} attribute? Ignored if \code{x} does not have this attribute}
}
\value{
data.table with columns for the coordinates x, y and z of the starting and
end point of each polygon edge
}
\description{
Special reader function for polygon edge output of voro++.
}
\examples{
random_unique_points <- unique(data.table::data.table(
  id = NA,
  x = runif(10, 0, 100000),
  y = runif(10, 0, 100000),
  z = runif(10, 0, 100)
))
random_unique_points$id <- 1:nrow(random_unique_points)

voro_output <- tessellate(random_unique_points, unit_scaling = c(0.001, 0.001, 1))

polygon_points <- read_polygon_edges(voro_output)

cut_surfaces <- cut_polygons(polygon_points, c(20, 40, 60))

cut_surfaces_sf <- cut_polygons_to_sf(cut_surfaces, crs = 25832)

\dontrun{
polygons_z_20 <- sf::st_geometry(cut_surfaces_sf[cut_surfaces_sf$z == 20, ])
plot(
  polygons_z_20,
  col = sf::sf.colors(10, categorical = TRUE)
)
}

}
